AiModelTemplate/requirements.txt
langchain
transformers
torch
numpy
pinecone-client
google-cloud-aiplatform
